less suck
Separated JS from the HTML, split JS into different files based on the functions, fixed the deck assignments (it gave Adlai the same deck for 9 games???), added a deck usage matrix, added the ability to edit player and deck names and delete them (must re-generate after editing), added tabs, created a folder for the CSS, added an Undo button to the bracket view, made it save data locally as changes are logged. That's probably it?